Updated some distance metrics.

+++ test/bct_test.cpp
@@ -60,7 +60,7 @@ std::vector<gsl_matrix*> bct_test::to_gsl(const NDArray m, int start) {
  */
 Matrix bct_test::from_gsl(const gsl_vector* gslv, int start) {
 	Matrix m = Matrix(1, gslv->size - start);
-	for (int i = start; i < gslv->size; i++) {
+	for (int i = start; i < (int)gslv->size; i++) {
 		m(0, i - start) = gsl_vector_get(gslv, i);
 	}
 	return m;
@@ -71,8 +71,8 @@ Matrix bct_test::from_gsl(const gsl_vector* gslv, int start) {
  */
 Matrix bct_test::from_gsl(const gsl_matrix* gslm, int start_row, int start_col) {
 	Matrix m = Matrix(gslm->size1 - start_row, gslm->size2 - start_col);
-	for (int i = start_row; i < gslm->size1; i++) {
-		for (int j = start_col; j < gslm->size2; j++) {
+	for (int i = start_row; i < (int)gslm->size1; i++) {
+		for (int j = start_col; j < (int)gslm->size2; j++) {
 			m(i - start_row, j - start_col) = gsl_matrix_get(gslm, i, j);
 		}
 	}
@@ -89,9 +89,9 @@ NDArray bct_test::from_gsl(const std::vector<gsl_matrix*> gslm, int start) {
 	dim_v(1) = gslm[start]->size2;
 	dim_v(2) = gslm.size() - start;
 	NDArray m(dim_v);
-	for (int k = start; k < gslm.size(); k++) {
-		for (int i = 0; i < gslm[start]->size1; i++) {
-			for (int j = 0; j < gslm[start]->size2; j++) {
+	for (int k = start; k < (int)gslm.size(); k++) {
+		for (int i = 0; i < (int)gslm[start]->size1; i++) {
+			for (int j = 0; j < (int)gslm[start]->size2; j++) {
 				m(i, j, k - start) = gsl_matrix_get(gslm[k], i, j);
 			}
 		}
